edit:seperated into kakao maps api utils file
